FEAT: Add support for transforms of meshes
Meshes can be transformed from one coordinate system to another
using affines. When transform is applied to a mesh, all other
transforms must be modified to project to the correct coordinate
system. Transforms should also be saved and loaded with meshes.

This commit modifies the Mesh class to add support for transforms.
Affine transforms can now be applied to a mesh and loaded/saved.