FIX: Fix a bug where loading FreeSurfer meshes with no CRAS
FreeSurfer meshes usually have a CRAS translation in the metadata
of the header. When the mesh is preprocessed with mris_convert
with the option --to-scanner, this field is removed and causes
an error on load.

This commit modifies the loading of FreeSurfer meshes to verify
if the CRAS filed exists instead of assuming it does.