FEAT: Add support for mesh segmentation
Each vertex of a mesh can be assigned a label. The list of all the
labels of the vertices of a mesh represent a segmentation of the
mesh.

This commit adds a new class that represents the segmentation of
a mesh. Support is also added to save and load a segmentation in
GifTI format.