REF: Rename segmentation labels to keys
The label id for each vertex in a segmentation is accessed using
the `labels` property. This makes it ambiguous whether it refers
to the table of labels or the labels of the vertices.

This commit changes the name of the property to `keys`. The labels
property will be added in a following commit to return the
metadata of the labels.