MINOR: Modify segmentation labels to return a dict
When accessing the labels of a segmentation, it is convenient to
get only one label, iterate over keys and labels, or iterate over
labels only. The current implementation allows iterating over
keys and labels only.

This commit modifies the labels property of the Segmentation class
to return a dict instead of an ItemsView. This way, the user can
choose over what to iterate.