FEAT: Add support for normals in meshes
In addition to vertices and triangles, the normals of the vertices
are often kept with meshes as they are expensive to compute. The
mesh class does not provide any facilities to handle these.

This commit adds support for normals in meshes. It also adds
support to save and load normals in GIfTI files.