FEAT: Add support for viewing vertex data and mesh information
The nimesh script allows segmentations to be viewed, but not
vertex data. In addition, no information can be obtained on a
mesh from the command line. This makes it difficult to use the
viewer as the user must already know the segmentations/vertex data
names.

This commit adds support for using vertex data to color meshes and
adds an option to print information about a mesh from the
command line.