FIX: Fix incorrect assumption when loading from freesurfer
The 'from_freesufer' helper function assumes a default coordinate
system of RAS which is incorrect. The coordinate system depends on
the image used to extract the surface.

This commit changes the default coordinate system to UNKNOWN. When
loading from a freesufer folder, the correct coordinate system is
extracted from the MRI image if possible.