FIX: Ignore transfroms from a space to itself
nimesh refuses to add transforms to the current space of the
mesh. This causes problems when loading from a GIfTI file as
they may contain an identity transformation to the space of
the mesh. For examples, HCP datasets.

This commit modifies how GIfTI files are loaded to ignore any
transfrom from space to itself.

Closes #2.