FIX: Fix bug that prevents nameless segmentations
Segmentations saved with nimesh store the name of the
segmentation in the metadata of the label array. However, when
the segmentation is loaded from a file created using other
software, this metadata is not present. This causes an
exception to be raised.

This commit modifies loading of segmentations so that if no name
is found it the file, the name is set to an empty string.