FEAT: Allow mesh vertices to be modified
The position of a mesh's vertices is supplied at initialization
time and cannot be modified. This makes it difficult to change
the vertices as all the other information must be copied.

This commit allows vertices to be changed as long as the number
of vertices of a mesh does not change.