FIX: Fix bug where nameless segmentations can not be loaded
When a segmentation is saved to a GIfTI using other software than
nimesh, segmentations (labels) do not have a name saved in the
metadata. This prevents these segmentations from being loaded
in nimesh.

This commit changes how segmentations are loaded to name
nameless segmentations as 'unknown' and warn the user.