FEAT: Add support to compute adjacency matrices
The adjacency matrix of a mesh provides information on the
topology of the mesh and is used by many algorithms. nimesh could
simplify these algorithms by generating the adjacecy matrix of
a Mesh.

This commit adds a function in nimesh.asarray to compute the
adjacency matrix of a mesh from its triangles. A convenience
property was also added to the Mesh class.