FIX: Fix bug caused by missing segmentations
The nimesh.io.freesurfer.load function expects the 3 common
segmentations files to be present in the FreeSurfer folder. When
they are not, the load function raises an exception. It should
instead warn the user and continue.

This commit tries to load the standard segmentations, but
continues without warning if they are missing. Warnings will be
added in upcoming changes.